Create AWS LambdaHandler
This code has been extracted from the Internet and I have not tested this code because of lack of Environment. I have my Company Sandbox but I do not have IAM permissions and Role to create a CMK key and write a Lambda handler to test the code. I believe that doing some adjustments and syntax check this should run fine.